+++ source/mclib/ablstd.cpp
@@ -52,53 +52,51 @@ bool   EnterStateSymbol = false;
 // STANDARD ROUTINE PARSING FUNCTIONS
 //***************************************************************************
 
-void stdReturn (void) {
-
+void stdReturn(void)
+{
 	// RETURN function
 	//
 	//		PARAMS: <same as function return type>
 	//
 	//		RETURN: NONE
-
-	if (curToken == TKN_LPAREN) {
+	if(curToken == TKN_LPAREN)
+	{
 		getToken();
 		TypePtr paramType = expression();
-		if (paramType != CurRoutineIdPtr->typePtr)
+		if(paramType != CurRoutineIdPtr->typePtr)
 			syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 		ifTokenGetElseError(TKN_RPAREN, ABL_ERR_SYNTAX_MISSING_RPAREN);
-		}
-	else if (CurRoutineIdPtr->typePtr != nullptr)
+	}
+	else if(CurRoutineIdPtr->typePtr != nullptr)
 		syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
 }
 
 //***************************************************************************
 
-void stdPrint (void) {
-
+void stdPrint(void)
+{
 	// PRINT function:
 	//
 	//		PARAMS:	integer or real or string
 	//
 	//		RETURN: NONE
-
-	if (curToken == TKN_LPAREN)
+	if(curToken == TKN_LPAREN)
 		getToken();
 	else
 		syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
-
 	TypePtr paramType = expression();
-	if ((paramType != IntegerTypePtr) &&
-		(paramType != RealTypePtr) &&
-		(paramType != CharTypePtr) &&
-		((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr)))
+	if((paramType != IntegerTypePtr) &&
+			(paramType != RealTypePtr) &&
+			(paramType != CharTypePtr) &&
+			((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr)))
 		syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 	ifTokenGetElseError(TKN_RPAREN, ABL_ERR_SYNTAX_MISSING_RPAREN);
 }
 
 //***************************************************************************
 
-TypePtr stdConcat (void) {
-
+TypePtr stdConcat(void)
+{
 	// PRINT function:
 	//
 	//		PARAMS:	char array
@@ -106,57 +104,50 @@ TypePtr stdConcat (void) {
 	//				integer, real or char array
 	//
 	//		RETURN: integer (resulting length, not including nullptr)
-
-	if (curToken == TKN_LPAREN)
+	if(curToken == TKN_LPAREN)
 		getToken();
 	else
 		syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
-
 	TypePtr paramType = expression();
-	if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
+	if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
 		syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 	ifTokenGetElseError(TKN_COMMA, ABL_ERR_SYNTAX_MISSING_COMMA);
-
 	paramType = expression();
-	if ((paramType != IntegerTypePtr) &&
-		(paramType != RealTypePtr) &&
-		(paramType != CharTypePtr) &&
-		(paramType != BooleanTypePtr) &&
-		((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr)))
+	if((paramType != IntegerTypePtr) &&
+			(paramType != RealTypePtr) &&
+			(paramType != CharTypePtr) &&
+			(paramType != BooleanTypePtr) &&
+			((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr)))
 		syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
-		
 	ifTokenGetElseError(TKN_RPAREN, ABL_ERR_SYNTAX_MISSING_RPAREN);
-
 	return(IntegerTypePtr);
 }
 
 //***************************************************************************
 
-TypePtr stdGetStateHandle (void) {
-
-	if (curToken == TKN_LPAREN)
+TypePtr stdGetStateHandle(void)
+{
+	if(curToken == TKN_LPAREN)
 		getToken();
 	else
 		syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
-
 	EnterStateSymbol = true;
 	TypePtr paramType = expression();
 	EnterStateSymbol = false;
-	if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
+	if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
 		syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
-
 	ifTokenGetElseError(TKN_RPAREN, ABL_ERR_SYNTAX_MISSING_RPAREN);
-
 	return(IntegerTypePtr);
 }
 
 //***************************************************************************
 
-TypePtr standardRoutineCall (SymTableNodePtr routineIdPtr) {
-
+TypePtr standardRoutineCall(SymTableNodePtr routineIdPtr)
+{
 	int32_t key = routineIdPtr->defn.info.routine.key;
 	int32_t numParams = FunctionInfoTable[key].numParams;
-	switch (key) {
+	switch(key)
+	{
 		case RTN_RETURN:
 			stdReturn();
 			return(nullptr);
@@ -168,66 +159,71 @@ TypePtr standardRoutineCall (SymTableNodePtr routineIdPtr) {
 		case RTN_GET_STATE_HANDLE:
 			return(stdGetStateHandle());
 		default:
-			if (key >= NumStandardFunctions)
+			if(key >= NumStandardFunctions)
 				syntaxError(ABL_ERR_SYNTAX_UNEXPECTED_TOKEN);
-			if (numParams == 0) {
-				if (curToken == TKN_LPAREN)
+			if(numParams == 0)
+			{
+				if(curToken == TKN_LPAREN)
 					syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
-				}
-			else {
-				if (curToken == TKN_LPAREN)
+			}
+			else
+			{
+				if(curToken == TKN_LPAREN)
 					getToken();
 				else
 					syntaxError(ABL_ERR_SYNTAX_WRONG_NUMBER_OF_PARAMS);
-				for (int32_t i = 0; i < numParams; i++) {
+				for(size_t i = 0; i < numParams; i++)
+				{
 					TypePtr paramType = expression();
-					switch (FunctionInfoTable[key].params[i]) {
+					switch(FunctionInfoTable[key].params[i])
+					{
 						case PARAM_TYPE_ANYTHING:
 							break;
 						case PARAM_TYPE_CHAR:
-							if (paramType != CharTypePtr)
+							if(paramType != CharTypePtr)
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_INTEGER:
-							if (paramType != IntegerTypePtr)
+							if(paramType != IntegerTypePtr)
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_REAL:
-							if (paramType != RealTypePtr)
+							if(paramType != RealTypePtr)
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_BOOLEAN:
-							if (paramType != BooleanTypePtr)
+							if(paramType != BooleanTypePtr)
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_INTEGER_REAL:
-							if ((paramType != IntegerTypePtr) && (paramType != RealTypePtr))
+							if((paramType != IntegerTypePtr) && (paramType != RealTypePtr))
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_CHAR_ARRAY:
-							if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
+							if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != CharTypePtr))
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_INTEGER_ARRAY:
-							if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != IntegerTypePtr))
+							if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != IntegerTypePtr))
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_REAL_ARRAY:
-							if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != RealTypePtr))
+							if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != RealTypePtr))
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 						case PARAM_TYPE_BOOLEAN_ARRAY:
-							if ((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != BooleanTypePtr))
+							if((paramType->form != FRM_ARRAY) || (paramType->info.array.elementTypePtr != BooleanTypePtr))
 								syntaxError(ABL_ERR_SYNTAX_INCOMPATIBLE_TYPES);
 							break;
 					}
-					if (i == (numParams - 1))
+					if(i == (numParams - 1))
 						ifTokenGetElseError(TKN_RPAREN, ABL_ERR_SYNTAX_MISSING_RPAREN);
 					else
 						ifTokenGetElseError(TKN_COMMA, ABL_ERR_SYNTAX_MISSING_COMMA);
 				}
 			}
-			switch (FunctionInfoTable[key].returnType) {
+			switch(FunctionInfoTable[key].returnType)
+			{
 				case RETURN_TYPE_NONE:
 					return(nullptr);
 				case RETURN_TYPE_INTEGER:
